refactor(cli): split the gde_cli god module into a commands package
Port of pro-v3-analysis e981212, with one addition core forced.

gde_cli.py was 1974 lines: the argument parser, ~20 command handlers, the
shared Rich formatting helpers and the dispatcher, in one file at fan_out 31 —
the last CRITICAL anti-pattern in this package.

Every one of its project imports was function-local and owned by exactly one
handler, so this is a relocation rather than an untangling: moving a handler
moved its imports with it, and functions moved verbatim.

    commands/parser.py             0   every subcommand, flag and help string
    commands/formatting.py         1   Rich output, plain-text fallbacks
    commands/voice_setup.py        3   provisioning + setup/check/speak
    commands/session_cmds.py       4   decide, recover, harden, explain
    commands/project_cmds.py       7   memory, ui, purge, doctor, engines, license
    commands/analysis_cmds.py      7   deps, advise, fetch, sync, telemetry
    commands/router.py             8   argv routing and dispatch
    commands/companion_cmds.py     8   the companion family
    commands/companion_backend.py  9   the startup --serve and --ui share
    gde_cli.py                     1   compatibility surface (132 lines)

Named `commands`, not `cli`: `genesis_architect.cli` is already the Typer app
for the scaffolding commands, and two sibling modules named `cli` in one
distribution is a readability hazard even though the namespaces do not
collide.

companion_backend.py is the addition. Relocation alone could not get the
companion family under the agreed ceiling of 11 — measured at 14, and every
pure-move seam landed at 11 or 12 because cmd_companion_ui carries nine
project modules on its own. The reason turned out to be real duplication:
`--serve` and `--ui` both brought the backend up with the same seven imports
and the same seven steps in the same order. Two copies of a startup sequence
is a live bug risk regardless of fan-out, since a fix to one can miss the
other. Extracting it takes the pair to 8 and 9.

Only the startup is shared. The two teardowns genuinely differ — `--serve`
uninstalls both patches and stops the IDE bridge, `--ui` does neither — and
that asymmetry is preserved rather than unified. It may well be a defect, but
deciding that is a behaviour change and this was a move.

One narrow behavioural difference is documented at the call site. The old
inline code assigned ws_server before the later steps, so a failure *after*
the socket opened left a live server with no router wired and `--ui` blocked
forever serving a backend that could not answer. start_backend() is atomic, so
a partial failure now takes the offline-UI path. Only the error path differs,
and only in that direction.

gde_cli.py stays because cli_entry.py imports main from it and that is what
pyproject declares as the console script; entry points resolve at install
time, so repointing the chain would force a reinstall everywhere for no
user-visible gain. Its 43 names resolve lazily (PEP 562), generated from the
package so the table is exhaustive by construction.

Verification, given the size. A snapshot tool walks the parser and dumps every
subcommand, flag, nargs, default, type, choices and help string — 413 lines,
captured from a worktree at the pre-refactor commit and diffed after every one
of the six steps. Byte-identical throughout. `ruff --select F821` ran after
each extraction, which is what caught the missing `sys`/`Path` in the Pro port
that the suite did not. Lint debt is conserved exactly: 19 I001 before, 19
after, zero introduced — the moved function-local import blocks carried their
own.

critical 1 -> 0, score 88 -> 89, cycles 0. Suite 2836 passed at every step.
